[issue #34] reactivate logging options
The Logging options panel is reactivated so that ABF users may now
set the log level at runtime directly from the options. In order to have
the options more prominent and to ensure that osx users can access the
options at all (on some systems the "Preferences" option is not shown
for an unknown reason) the "ABF" menu is introduced.