[issue #34] new domainserver options
Adds new options panel to globally control domainserver properties. The
first and currently only property to configure is the refresh policy
that is the scope of this issue. While the option is now available it
still has to be used at the appripriate refresh actions so that the
configuration actually has an effect.